refactor & safer string formatting
