+++ base/oauth.go
@@ -143,6 +143,10 @@ type GetOAuthOpts struct {
 	OAuthOfflineAccessType bool
 	// if the token is not found, don't request new authorization (default: false)
 	SkipAuthentication bool
+	// template for the auth message (default: "Visit %s\n to authorize me.")
+	AuthMessageTemplate string
+	// optional callback which constructs and sends auth URL (default: disabled)
+	AuthURLCallback func(authUrl string) error
 }
 
 func GetOAuthClient(
@@ -152,7 +156,6 @@ func GetOAuthClient(
 	requests *OAuthRequests,
 	config *oauth2.Config,
 	storage OAuthStorage,
-	authMessageTemplate string,
 	opts GetOAuthOpts,
 ) (*http.Client, error) {
 	token, err := storage.GetToken(tokenIdentifier)
@@ -192,16 +195,30 @@ func GetOAuthClient(
 		authURL := config.AuthCodeURL(state, oauthOpts...)
 		// strip protocol to skip unfurl prompt
 		authURL = strings.TrimPrefix(authURL, "https://")
-		_, err = kbc.SendMessageByTlfName(callbackMsg.Sender.Username, authMessageTemplate, authURL)
+		if opts.AuthURLCallback != nil {
+			err = opts.AuthURLCallback(authURL)
+		} else {
+			authMessageTemplate := opts.AuthMessageTemplate
+			if authMessageTemplate == "" {
+				authMessageTemplate = "Visit %s\n to authorize me."
+			}
+			_, err = kbc.SendMessageByTlfName(callbackMsg.Sender.Username, authMessageTemplate, authURL)
+		}
+		if err != nil {
+			return nil, fmt.Errorf("error sending message: %s", err)
+		}
 
 		// If we are in a 1-1 conv directly or as a bot user with the sender, skip this message.
 		if callbackMsg.Channel.MembersType == "team" || !(callbackMsg.Sender.Username == callbackMsg.Channel.Name ||
 			len(strings.Split(callbackMsg.Channel.Name, ",")) == 2) {
 			_, err = kbc.SendMessageByConvID(callbackMsg.ConvID,
 				"OK! I've sent a message to @%s to authorize me.", callbackMsg.Sender.Username)
+			if err != nil {
+				return nil, fmt.Errorf("error sending message: %s", err)
+			}
 		}
 
-		return nil, err
+		return nil, nil
 	}
 
 	return config.Client(context.Background(), token), nil
