+++ gcalbot/gcalbot/accounts.go
@@ -2,10 +2,38 @@ package gcalbot
 
 import (
 	"fmt"
+	"strings"
+
+	"github.com/kballard/go-shellquote"
 
 	"github.com/keybase/go-keybase-chat-bot/kbchat/types/chat1"
 	"github.com/keybase/managed-bots/base"
 )
+
+func (h *Handler) HandleAuth(msg chat1.MsgSummary) error {
+	cmd := strings.TrimSpace(msg.Content.Text.Body)
+	toks, err := shellquote.Split(cmd)
+	if err != nil {
+		h.ChatDebug(msg.ConvID, "error splitting command: %s", err)
+		return nil
+	}
+	args := toks[3:]
+	if len(args) != 1 {
+		h.ChatDebug(msg.ConvID, "bad args for accounts connect: %s", args)
+		return nil
+	}
+
+	username := msg.Sender.Username
+	accountNickname := args[0]
+	err = h.db.InsertAccountForUser(username, accountNickname)
+	if err != nil {
+		h.ChatDebug(msg.ConvID, "error connecting account %s", accountNickname)
+		return err
+	}
+
+	_, err = h.kbc.SendMessageByConvID(msg.ConvID, "Account '%s' has been connected.", accountNickname)
+	return err
+}
 
 func (h *Handler) accountsListHandler(msg chat1.MsgSummary) error {
 	username := msg.Sender.Username
@@ -23,11 +51,8 @@ func (h *Handler) accountsListHandler(msg chat1.MsgSummary) error {
 		return nil
 	}
 
-	accountListMessage := "Here are your connected accounts:"
-	for index := range accounts {
-		accountListMessage += fmt.Sprintf("\n• %s", accounts[index])
-	}
-	_, err = h.kbc.SendMessageByConvID(msg.ConvID, accountListMessage)
+	accountListMessage := "Here are your connected accounts:" + strings.Repeat("\n• %s", len(accounts))
+	_, err = h.kbc.SendMessageByConvID(msg.ConvID, accountListMessage, accounts...)
 	if err != nil {
 		return fmt.Errorf("error sending message: %s", err)
 	}
@@ -65,11 +90,17 @@ func (h *Handler) accountsConnectHandler(cmd string, msg chat1.MsgSummary) error
 
 	// TODO(marcel): is this the best method? adding more columns to the db complicates the existing oauth abstractions
 	identifier := fmt.Sprintf("%s:%s", username, accountNickname)
-	// TODO(marcel): this is very hacky
-	authMessageTemplate := fmt.Sprintf("Visit %s to connect a Google account as '%s'.", "%s", accountNickname)
 
+	authURLCallback := func(authURL string) error {
+		_, err = h.kbc.SendMessageByTlfName(msg.Sender.Username,
+			"Visit %s to connect a Google account as '%s'.", authURL, accountNickname)
+		return err
+	}
 	_, err = base.GetOAuthClient(identifier, msg, h.kbc, h.requests, h.config, h.db,
-		authMessageTemplate, base.GetOAuthOpts{OAuthOfflineAccessType: true})
+		base.GetOAuthOpts{
+			OAuthOfflineAccessType: true,
+			AuthURLCallback:        authURLCallback,
+		})
 	if err != nil {
 		return fmt.Errorf("error authenticating user: %s", err)
 	}
