+++ gcalbot/gcalbot/handler.go
@@ -3,8 +3,6 @@ package gcalbot
 import (
 	"strings"
 
-	"github.com/kballard/go-shellquote"
-
 	"github.com/keybase/go-keybase-chat-bot/kbchat"
 	"github.com/keybase/go-keybase-chat-bot/kbchat/types/chat1"
 	"github.com/keybase/managed-bots/base"
@@ -35,31 +33,6 @@ func NewHandler(kbc *kbchat.API, db *DB, requests *base.OAuthRequests, config *o
 func (h *Handler) HandleNewConv(conv chat1.ConvSummary) error {
 	welcomeMsg := "Hello! I can get you setup with a Google Meet video call anytime, just send me `!meet`."
 	return base.HandleNewTeam(h.DebugOutput, h.kbc, conv, welcomeMsg)
-}
-
-func (h *Handler) HandleAuth(msg chat1.MsgSummary) error {
-	cmd := strings.TrimSpace(msg.Content.Text.Body)
-	toks, err := shellquote.Split(cmd)
-	if err != nil {
-		h.ChatDebug(msg.ConvID, "error splitting command: %s", err)
-		return nil
-	}
-	args := toks[3:]
-	if len(args) != 1 {
-		h.ChatDebug(msg.ConvID, "bad args for accounts connect: %s", args)
-		return nil
-	}
-
-	username := msg.Sender.Username
-	accountNickname := args[0]
-	err = h.db.InsertAccountForUser(username, accountNickname)
-	if err != nil {
-		h.ChatDebug(msg.ConvID, "error connecting account %s", accountNickname)
-		return err
-	}
-
-	_, err = h.kbc.SendMessageByConvID(msg.ConvID, "Account '%s' has been connected.", accountNickname)
-	return err
 }
 
 func (h *Handler) HandleCommand(msg chat1.MsgSummary) error {
