+++ meetbot/meetbot/handler.go
@@ -74,8 +74,10 @@ func (h *Handler) meetHandler(msg chat1.MsgSummary) error {
 func (h *Handler) meetHandlerInner(msg chat1.MsgSummary) error {
 	identifier := base.IdentifierFromMsg(msg)
 	client, err := base.GetOAuthClient(identifier, msg, h.kbc, h.requests, h.config, h.db,
-		"Visit %s\n to authorize me to create events.",
-		base.GetOAuthOpts{OAuthOfflineAccessType: true})
+		base.GetOAuthOpts{
+			AuthMessageTemplate:    "Visit %s\n to authorize me to create events.",
+			OAuthOfflineAccessType: true,
+		})
 	if err != nil || client == nil {
 		return err
 	}
